made the super-user bar an area so it can be edited
The super user bar dialog uses a multi-value composite field to get a title and url for each menu item.
Other magnolia users have commented that this type of field is not working when configured by file, so
it is currently configured in the jcr.  We can make a yaml file once the problem is fixed.

The code that generates the menu based on jcr data is commented out right now.