Got started on mail stuff.
Added a mail area to gato-template and a mail template to the txstate2015 template that uses the standard template. The only difference between the standard template and mail template is the contentParagraph. Since the areas are shared between standard-template and mail-template I moved the areas config to a separate yaml file that can be shared.