Remove templatecomponent macro.
Talked with Nick about this and he wants to abandon the use of the templatecomponent macro inside of components since they make every component depend on gato-template. If we do want to place a wrapper around each component, that can still be done in the area and makes it possible to have a more useful, area specific wrapper around our components.